@@ -1,37 +1,43 @@
 import 'package:flutter/material.dart';
 import 'package:iconsax/iconsax.dart';
-import 'package:puntos_smart_user/app/core/constants/app_images.dart';
 import 'package:puntos_smart_user/app/core/constants/app_text.dart';
 import 'package:puntos_smart_user/app/core/theme/app_colors.dart';
 
 class ProfileModelItem {
-  final String image;
+  //final String? image;
   final String title;
   final Widget widget;
+  final IconData icon;
 
   ProfileModelItem(
-      {required this.image, required this.title, required this.widget});
+      {/*this.image,*/ required this.title,
+      required this.widget,
+      required this.icon});
 }
 
 final List<ProfileModelItem> poinstAndSubscription = [
   ProfileModelItem(
-    image: AppImages.wallet,
+    //image: AppImages.wallet,
+    icon: Iconsax.wallet_money,
     title: AppText.myWallet,
     widget: Container(
       decoration: BoxDecoration(
         color: AppColors.onPrimary,
-        borderRadius: BorderRadius.circular(15),
+        borderRadius: BorderRadius.circular(999),
       ),
       child: const Padding(
         padding: EdgeInsets.all(8.0),
         child: Text(
           "50 ps",
+          style:
+              TextStyle(color: AppColors.primary, fontWeight: FontWeight.w800),
         ),
       ),
     ),
   ),
   ProfileModelItem(
-    image: AppImages.coins,
+    //image: AppImages.coins,
+    icon: Iconsax.money_change,
     title: AppText.sendPoints,
     widget: const Icon(
       Icons.navigate_next,
@@ -39,7 +45,8 @@ final List<ProfileModelItem> poinstAndSubscription = [
     ),
   ),
   ProfileModelItem(
-    image: AppImages.fire,
+    //image: AppImages.fire,
+    icon: Iconsax.crown,
     title: AppText.subcriptionSmart,
     widget: const Icon(
       Icons.navigate_next,
@@ -49,7 +56,8 @@ final List<ProfileModelItem> poinstAndSubscription = [
 ];
 final List<ProfileModelItem> profileList = [
   ProfileModelItem(
-    image: AppImages.fire,
+    //image: AppImages.fire,
+    icon: Iconsax.profile,
     title: AppText.personalInformation,
     widget: const Icon(
       Iconsax.user,
@@ -57,34 +65,38 @@ final List<ProfileModelItem> profileList = [
     ),
   ),
   ProfileModelItem(
-    image: AppImages.fire,
+    //image: AppImages.fire,
+    icon: Iconsax.map,
     title: AppText.address,
     widget: const Icon(
-      Icons.directions,
+      Iconsax.map,
       color: AppColors.onPrimary,
     ),
   ),
   ProfileModelItem(
-    image: AppImages.fire,
+    //image: AppImages.fire,
+    icon: Iconsax.like_dislike,
     title: AppText.personalPreferences,
     widget: const Icon(
-      Icons.room_preferences,
+      Iconsax.like_dislike,
       color: AppColors.onPrimary,
     ),
   ),
-  ProfileModelItem(
-    image: AppImages.fire,
-    title: AppText.myFavorities,
-    widget: const Icon(
-      Icons.favorite,
-      color: AppColors.onPrimary,
-    ),
-  ),
+  // ProfileModelItem(
+  //   //image: AppImages.fire,
+  //   icon: Iconsax.heart,
+  //   title: AppText.myFavorities,
+  //   widget: const Icon(
+  //     Iconsax.heart4, //5
+  //     color: AppColors.onPrimary,
+  //   ),
+  // ),
 ];
 
 final List<ProfileModelItem> winPontsListProfile = [
   ProfileModelItem(
-    image: AppImages.coins,
+    //image: AppImages.coins,
+    icon: Iconsax.game,
     title: AppText.gameSmart,
     widget: const Icon(
       Icons.navigate_next,
@@ -92,7 +104,8 @@ final List<ProfileModelItem> winPontsListProfile = [
     ),
   ),
   ProfileModelItem(
-    image: AppImages.fire,
+    //image: AppImages.fire,
+    icon: Iconsax.task_square,
     title: AppText.answerAndWind,
     widget: const Icon(
       Icons.navigate_next,
@@ -103,7 +116,8 @@ final List<ProfileModelItem> winPontsListProfile = [
 
 final List<ProfileModelItem> smartSupportList = [
   ProfileModelItem(
-    image: AppImages.chat,
+    //image: AppImages.chat,
+    icon: Iconsax.send_2,
     title: AppText.chatSmart,
     widget: const Icon(
       Icons.navigate_next,
@@ -111,7 +125,8 @@ final List<ProfileModelItem> smartSupportList = [
     ),
   ),
   ProfileModelItem(
-    image: AppImages.coments,
+    //image: AppImages.coments,
+    icon: Iconsax.message_search,
     title: AppText.faq,
     widget: const Icon(
       Icons.navigate_next,
@@ -119,7 +134,8 @@ final List<ProfileModelItem> smartSupportList = [
     ),
   ),
   ProfileModelItem(
-    image: AppImages.phoneAlert,
+    //image: AppImages.phoneAlert,
+    icon: Iconsax.call,
     title: AppText.supportServices,
     widget: const Icon(
       Icons.navigate_next,
@@ -127,7 +143,8 @@ final List<ProfileModelItem> smartSupportList = [
     ),
   ),
   ProfileModelItem(
-    image: AppImages.termns,
+    //image: AppImages.termns,
+    icon: Iconsax.document_text,
     title: AppText.termsAndContion,
     widget: const Icon(
       Icons.navigate_next,
@@ -135,7 +152,8 @@ final List<ProfileModelItem> smartSupportList = [
     ),
   ),
   ProfileModelItem(
-    image: AppImages.termns,
+    //image: AppImages.termns,
+    icon: Iconsax.document_text,
     title: AppText.privacyPolicy,
     widget: const Icon(
       Icons.navigate_next,
