@@ -1,5 +1,6 @@
 import 'dart:math' as math;
 import 'package:flutter/material.dart';
+import 'package:puntos_smart_user/app/core/extension/custom_extension.dart';
 import 'package:puntos_smart_user/app/features/coupon_feature/presentation/painters/coupon_painter.dart';
 
 import '../../../../core/theme/app_colors.dart';
@@ -47,61 +48,101 @@ class CouponWidget extends StatelessWidget {
           ),
           Positioned(
             left: 20,
-            child: Container(
-              height: 100,
+            child: SizedBox(
               width: MediaQuery.of(context).size.width * 0.6,
-              alignment: Alignment.center,
-              child: Row(
+              child: Column(
                 children: [
-                  Expanded(
-                    flex: 7, // 70% del ancho para el texto
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      crossAxisAlignment: CrossAxisAlignment.start,
+                  Container(
+                    alignment: Alignment.center,
+                    child: Row(
                       children: [
-                        Text(
-                          'Titulo Descuento en la compra de la segunda hamburguesa',
-                          style: textTheme.titleSmall!.copyWith(
-                            color: Colors.black54,
-                            fontWeight: FontWeight.bold,
-                            fontFamily: 'onset',
-                            letterSpacing: 1,
-                            overflow: TextOverflow.ellipsis,
-                            height: 1.1,
+                        Expanded(
+                          flex: 7, // 70% del ancho para el texto
+                          child: Column(
+                            mainAxisAlignment: MainAxisAlignment.center,
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Text(
+                                'Titulo Descuento en la compra de la segunda hamburguesa',
+                                style: textTheme.titleSmall!.copyWith(
+                                  color: AppColors.primary,
+                                  fontWeight: FontWeight.w600,
+                                  fontFamily: 'onset',
+                                  letterSpacing: 1,
+                                  overflow: TextOverflow.ellipsis,
+                                  height: 1.1,
+                                  fontSize: 16,
+                                ),
+                                maxLines: 2,
+                              ),
+                              Text(
+                                'Descripcion Descuento en la compra de la segunda hamburguesa',
+                                style: textTheme.labelMedium!.copyWith(
+                                  color: AppColors.primary,
+                                  //fontWeight: FontWeight.w400,
+                                  //fontFamily: 'onset',
+                                  letterSpacing: 1,
+                                  overflow: TextOverflow.ellipsis,
+                                  height: 1.1,
+                                ),
+                                maxLines: 2,
+                              ),
+                            ],
                           ),
-                          maxLines: 2,
                         ),
-                        Text(
-                          'Descripcion Descuento en la compra de la segunda hamburguesa',
-                          style: textTheme.labelMedium!.copyWith(
-                            color: Colors.black54,
-                            fontWeight: FontWeight.bold,
-                            fontFamily: 'onset',
-                            letterSpacing: 1,
-                            overflow: TextOverflow.ellipsis,
-                            height: 1.1,
+                        Flexible(
+                          flex: 3, // 30% del ancho para la imagen
+                          child: Container(
+                            width: double
+                                .infinity, // Ocupa todo el espacio disponible (30% del ancho)
+                            height: 80,
+                            decoration: const BoxDecoration(
+                              shape: BoxShape
+                                  .circle, // Hace que el contenedor sea circular
+                              image: DecorationImage(
+                                image: AssetImage(
+                                    'assets/images/hamburguesa2.jpg'), // Reemplaza con la ruta de tu imagen
+                                fit: BoxFit
+                                    .cover, // Ajusta la imagen para cubrir el contenedor
+                              ),
+                            ),
                           ),
-                          maxLines: 2,
                         ),
                       ],
                     ),
                   ),
-                  Flexible(
-                    flex: 3, // 30% del ancho para la imagen
-                    child: Container(
-                      width: double
-                          .infinity, // Ocupa todo el espacio disponible (30% del ancho)
-                      height: 80,
-                      decoration: const BoxDecoration(
-                        shape: BoxShape
-                            .circle, // Hace que el contenedor sea circular
-                        image: DecorationImage(
-                          image: AssetImage(
-                              'assets/images/hamburguesa2.jpg'), // Reemplaza con la ruta de tu imagen
-                          fit: BoxFit
-                              .cover, // Ajusta la imagen para cubrir el contenedor
-                        ),
+                  8.hSpace,
+                  SizedBox(
+                    width: double
+                        .infinity, // Ocupa el 100% del ancho del contenedor
+                    child: Text(
+                      '45PS',
+                      style: textTheme.titleLarge!.copyWith(
+                        color: AppColors.primary,
+                        fontWeight: FontWeight.bold,
+                        fontFamily: 'poppins',
+                        letterSpacing: 1,
+                        overflow: TextOverflow.ellipsis,
+                        height: 1.1,
                       ),
+                      maxLines: 1,
+                    ),
+                  ),
+                  3.hSpace,
+                  SizedBox(
+                    width: double
+                        .infinity, // Ocupa el 100% del ancho del contenedor
+                    child: Text(
+                      'Valido de 15/10/2024 hasta 28/10/2024',
+                      style: textTheme.titleSmall!.copyWith(
+                        color: AppColors.primary,
+                        //fontWeight: FontWeight.w500,
+                        letterSpacing: 1,
+                        overflow: TextOverflow.ellipsis,
+                        height: 1.1,
+                        fontSize: 8,
+                      ),
+                      maxLines: 2,
                     ),
                   ),
                 ],
@@ -120,8 +161,8 @@ class CouponWidget extends StatelessWidget {
                       text: '35 % \n', // Parte del texto con un estilo
                       style: textTheme.titleLarge!.copyWith(
                         color: AppColors.primary,
-                        fontWeight: FontWeight.w900,
-                        fontFamily: 'onest',
+                        fontWeight: FontWeight.bold,
+                        fontFamily: 'poppins',
                       ),
                     ),
                     TextSpan(
@@ -142,26 +183,3 @@ class CouponWidget extends StatelessWidget {
     );
   }
 }
-
-
-// const SizedBox(height: 6),
-                        // Text(
-                        //   '45 PS',
-                        //   style: textTheme.titleMedium!.copyWith(
-                        //     color: Colors.black87,
-                        //     fontWeight: FontWeight.w900,
-                        //     fontFamily: 'onest',
-                        //     letterSpacing: 2,
-                        //   ),
-                        // ),
-                        // const SizedBox(height: 4),
-                        // Text(
-                        //   '15/10/2024 hasta 28/10/2024',
-                        //   style: textTheme.labelSmall!.copyWith(
-                        //     color: Colors.black54,
-                        //     fontWeight: FontWeight.w500,
-                        //     fontFamily: 'onest',
-                        //     letterSpacing: 1,
-                        //   ),
-                        //   textAlign: TextAlign.right,
-                        // ),
